third_party/blink/web_tests/fast/hidpi/2x/image-srcset-invalid-inputs-except-one.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }

    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==200');
    }, false);
</script>
<div>This test passes if the img tag below is a green square regardless of the scale factor. It ensures that invalid inputs are
    ignored and well-formed images are chosen regardless of their qualifiers, once they are the only candidate left</div>
<img id="foo" src="" srcset="1x,,  ,   x    ,2x  , foo.jpg, 3x, bar.jpg 4x 100h, foo.jpg 5, bar.jpg dx, ../resources/green-400-px-square.png   2x ,">
